Improve IOLoop timeout fast path
Remove the size limit for the timeout queue. Instead, check
both ends of the heap. Heaps are *roughly* sorted by deadline,
so if both ends are due, it's likely the optimization would
kick in, and we don't run a huge risk of iterating through
the list unnecessarily.

That also makes the optimization more useful, since it can
handle huge lists of due timeouts. Take the opportunity to
maintain timeout order by using an in-place sort. The algorithm
for in-place sort should still be better than extracting them
one at a time, which would amount to a rather inefficient
implementation of heap sort.